refactor: 修改认证服务结构
