+++ pom.xml
@@ -11,7 +11,7 @@
         <module>taroco-docs</module>
         <module>taroco-cloud</module>
         <module>taroco-gateway</module>
-        <module>taroco-oauth2</module>
+        <module>taroco-authentication</module>
         <module>taroco-rbac</module>
     </modules>
     <packaging>pom</packaging>
@@ -100,11 +100,6 @@
             <dependency>
                 <groupId>cn.taroco</groupId>
                 <artifactId>taroco-swagger2-spring-boot-starter</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>cn.taroco</groupId>
-                <artifactId>oauth2-config</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <!-- spring-boot-admin server -->
