+++ taroco-authentication/pom.xml
@@ -3,19 +3,19 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>taroco-oauth2</artifactId>
+        <artifactId>taroco</artifactId>
         <groupId>cn.taroco</groupId>
         <version>2.0.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>authentication-server</artifactId>
+    <artifactId>taroco-authentication</artifactId>
     <packaging>jar</packaging>
-    <name>authentication-server</name>
+    <name>taroco-authentication</name>
 
     <dependencies>
         <dependency>
-            <groupId>cn.taroco</groupId>
-            <artifactId>oauth2-config</artifactId>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-starter-oauth2</artifactId>
         </dependency>
         <dependency>
             <groupId>cn.taroco</groupId>
