+++ taroco-authentication/src/main/java/cn/taroco/oauth2/server/TarocoOauth2ServerApplication.java
@@ -1,6 +1,5 @@
 package cn.taroco.oauth2.server;
 
-import cn.taroco.oauth2.config.annotation.EnableAuthJwtTokenStore;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
@@ -15,7 +14,6 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
  */
 @SpringBootApplication
 @EnableDiscoveryClient
-@EnableAuthJwtTokenStore
 @EnableFeignClients
 public class TarocoOauth2ServerApplication {
