+++ taroco-authentication/src/main/java/cn/taroco/oauth2/server/config/SecurityMessageLocal.java
