+++ taroco-authentication/src/main/java/cn/taroco/oauth2/server/controller/AuthenticationController.java
