+++ taroco-authentication/src/main/resources/bootstrap.yml
@@ -1,12 +1,6 @@
 server:
   port: 9001
 
-encrypt:
-  key-store:
-    location: classpath:taroco.jks
-    secret: taroco!@#$
-    alias: taroco
-
 spring:
   application:
     name: taroco-authentication-server
