+++ taroco-authentication/src/main/resources/pubkey.txt
