+++ taroco-authentication/src/main/resources/static/css/common.css
