+++ taroco-authentication/src/main/resources/static/css/style.css
