+++ taroco-authentication/src/main/resources/static/fonts/fontawesome-webfont.woff2
