+++ taroco-authentication/src/main/resources/static/js/common.js
