+++ taroco-authentication/src/main/resources/taroco.jks
