+++ taroco-gateway/pom.xml
@@ -14,10 +14,6 @@
     <dependencies>
         <dependency>
             <groupId>cn.taroco</groupId>
-            <artifactId>oauth2-config</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>cn.taroco</groupId>
             <artifactId>taroco-common-spring-boot-starter</artifactId>
         </dependency>
         <dependency>
