+++ taroco-gateway/src/main/java/cn/taroco/gateway/TarocoGatewayApplication.java
@@ -1,6 +1,5 @@
 package cn.taroco.gateway;
 
-import cn.taroco.oauth2.config.annotation.EnableResJwtTokenStore;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
@@ -16,7 +15,6 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 @EnableDiscoveryClient
 @SpringBootApplication
 @EnableFeignClients
-@EnableResJwtTokenStore
 public class TarocoGatewayApplication {
 
 	public static void main(String[] args) {
