+++ taroco-gateway/src/main/java/cn/taroco/gateway/config/ResourceServerConfiguration.java
@@ -1,12 +1,18 @@
 package cn.taroco.gateway.config;
 
+import cn.taroco.common.config.TarocoOauth2Properties;
 import cn.taroco.gateway.handler.TarocoAccessDeniedHandler;
-import cn.taroco.oauth2.config.FilterIgnorePropertiesConfig;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.security.oauth2.resource.ResourceServerProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.core.io.Resource;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configurers.ExpressionUrlAuthorizationConfigurer;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -15,6 +21,17 @@ import org.springframework.security.oauth2.config.annotation.web.configuration.E
 import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
 import org.springframework.security.oauth2.config.annotation.web.configurers.ResourceServerSecurityConfigurer;
 import org.springframework.security.oauth2.provider.expression.OAuth2WebSecurityExpressionHandler;
+import org.springframework.security.oauth2.provider.token.TokenStore;
+import org.springframework.security.oauth2.provider.token.store.JwtAccessTokenConverter;
+import org.springframework.security.oauth2.provider.token.store.JwtTokenStore;
+import org.springframework.web.client.RestTemplate;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.util.Base64;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 /**
  * 资源服务器配置
@@ -24,11 +41,13 @@ import org.springframework.security.oauth2.provider.expression.OAuth2WebSecurity
  */
 @Configuration
 @EnableResourceServer
-@EnableConfigurationProperties(FilterIgnorePropertiesConfig.class)
+@EnableConfigurationProperties(TarocoOauth2Properties.class)
 public class ResourceServerConfiguration extends ResourceServerConfigurerAdapter {
 
+    private static final String PUBLIC_KEY = "pubkey.txt";
+
     @Autowired
-    private FilterIgnorePropertiesConfig filterIgnorePropertiesConfig;
+    private TarocoOauth2Properties oauth2Properties;
 
     @Autowired
     private OAuth2WebSecurityExpressionHandler expressionHandler;
@@ -36,13 +55,61 @@ public class ResourceServerConfiguration extends ResourceServerConfigurerAdapter
     @Autowired
     private TarocoAccessDeniedHandler accessDeniedHandler;
 
+    @Autowired
+    private ResourceServerProperties resource;
+
+    @Bean
+    public TokenStore tokenStore(JwtAccessTokenConverter jwtAccessTokenConverter) {
+        return new JwtTokenStore(jwtAccessTokenConverter);
+    }
+
+    @Bean
+    public JwtAccessTokenConverter jwtAccessTokenConverter() {
+        JwtAccessTokenConverter converter = new JwtAccessTokenConverter();
+        converter.setVerifierKey(getPubKey());
+        return converter;
+    }
+
+    /**
+     * 获取非对称加密公钥 Key
+     * @return 公钥 Key
+     */
+    private String getPubKey() {
+        Resource resource = new ClassPathResource(PUBLIC_KEY);
+        try (BufferedReader br = new BufferedReader(new InputStreamReader(resource.getInputStream()))) {
+            return br.lines().collect(Collectors.joining("\n"));
+        } catch (IOException ioe) {
+            return getKeyFromAuthorizationServer();
+        }
+    }
+
+    /**
+     * 通过访问授权服务器获取非对称加密公钥 Key
+     * @return 公钥 Key
+     */
+    private String getKeyFromAuthorizationServer() {
+        final RestTemplate keyUriRestTemplate = new RestTemplate();
+        final HttpHeaders headers = new HttpHeaders();
+        final String username = this.resource.getClientId();
+        final String password = this.resource.getClientSecret();
+        if (username != null && password != null) {
+            final byte[] token = Base64.getEncoder().encode((username + ":" + password).getBytes());
+            headers.add("Authorization", "Basic " + new String(token));
+        }
+        final HttpEntity<Void> request = new HttpEntity<>(headers);
+        final String url = this.resource.getJwt().getKeyUri();
+        return (String) keyUriRestTemplate
+                .exchange(url, HttpMethod.GET, request, Map.class).getBody()
+                .get("value");
+    }
+
     @Override
     public void configure(HttpSecurity http) throws Exception {
         //允许使用iframe 嵌套，避免swagger-ui 不被加载的问题
         http.headers().frameOptions().disable();
         ExpressionUrlAuthorizationConfigurer<HttpSecurity>.ExpressionInterceptUrlRegistry registry = http
                 .authorizeRequests();
-        filterIgnorePropertiesConfig.getUrls().forEach(url -> registry.antMatchers(url).permitAll());
+        oauth2Properties.getUrlPermitAll().forEach(url -> registry.antMatchers(url).permitAll());
         registry.anyRequest()
                 .access("@permissionService.hasPermission(request, authentication)");
     }
