--- taroco-oauth2/oauth2-config/src/main/java/cn/taroco/oauth2/config/TarocoJwtAccessTokenConverter.java
@@ -1,34 +1,0 @@
-package cn.taroco.oauth2.config;
-
-import cn.taroco.common.constants.SecurityConstants;
-import org.springframework.security.oauth2.common.OAuth2AccessToken;
-import org.springframework.security.oauth2.provider.OAuth2Authentication;
-import org.springframework.security.oauth2.provider.token.store.JwtAccessTokenConverter;
-
-import java.util.Map;
-
-/**
- * @author liuht
- * @date 2018/4/7
- * token 声明版权
- */
-@SuppressWarnings("unchecked")
-public class TarocoJwtAccessTokenConverter extends JwtAccessTokenConverter {
-
-    @Override
-    public Map<String, ?> convertAccessToken(OAuth2AccessToken token, OAuth2Authentication authentication) {
-        Map<String, Object> representation = (Map<String, Object>) super.convertAccessToken(token, authentication);
-        representation.put("license", SecurityConstants.LICENSE);
-        return representation;
-    }
-
-    @Override
-    public OAuth2AccessToken extractAccessToken(String value, Map<String, ?> map) {
-        return super.extractAccessToken(value, map);
-    }
-
-    @Override
-    public OAuth2Authentication extractAuthentication(Map<String, ?> map) {
-        return super.extractAuthentication(map);
-    }
-}
