--- taroco-oauth2/oauth2-config/src/main/java/cn/taroco/oauth2/config/constants/AuthoritiesEnum.java
@@ -1,33 +1,0 @@
-package cn.taroco.oauth2.config.constants;
-
-/**
- * 权限常量
- *
- * @author liuht
- * @date 2018/7/24 16:05
- */
-public enum AuthoritiesEnum {
-    /**
-     * 管理员
-     */
-    ADMIN("ROLE_ADMIN"),
-    /**
-     * 普通用户
-     */
-    USER("ROLE_USER"),
-    /**
-     * 匿名用户
-     */
-    ANONYMOUS("ROLE_ANONYMOUS");
-
-    private String role;
-
-    AuthoritiesEnum(String role) {
-        this.role = role;
-    }
-
-    public String getRole() {
-        return role;
-    }
-
-}
