--- taroco-oauth2/pom.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>taroco</artifactId>
-        <groupId>cn.taroco</groupId>
-        <version>2.0.1-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-    <packaging>pom</packaging>
-    <modules>
-        <module>authentication-server</module>
-        <module>oauth2-config</module>
-    </modules>
-    <artifactId>taroco-oauth2</artifactId>
-
-    <dependencies>
-        <dependency>
-            <groupId>cn.taroco</groupId>
-            <artifactId>taroco-log-spring-boot-starter</artifactId>
-        </dependency>
-    </dependencies>
-</project>
